+++ pgvector/django/vector.py
@@ -1,6 +1,5 @@
 from django import forms
 from django.db.models import Field
-import numpy as np
 from typing import Any
 from .. import Vector
 
@@ -25,29 +24,22 @@ class VectorField(Field):
             return 'vector'
         return 'vector(%d)' % self.dimensions
 
-    def from_db_value(self, value: Any, expression: Any, connection: Any) -> np.ndarray | None:
+    def from_db_value(self, value: Any, expression: Any, connection: Any) -> Vector | None:
         return Vector._from_db(value)
 
-    def to_python(self, value: Any) -> np.ndarray | None:
-        if isinstance(value, list):
-            return np.array(value, dtype=np.float32)
-        return Vector._from_db(value)
+    def to_python(self, value: Any) -> Vector | None:
+        if value is None or isinstance(value, Vector):
+            return value
+        elif isinstance(value, str):
+            return Vector._from_db(value)
+        else:
+            return Vector(value)
 
     def get_prep_value(self, value: Any) -> str | None:
         return Vector._to_db(value)
 
     def value_to_string(self, obj: Any) -> str | None:
         return self.get_prep_value(self.value_from_object(obj))
-
-    def validate(self, value: Any, model_instance: Any) -> None:
-        if isinstance(value, np.ndarray):
-            value = value.tolist()
-        super().validate(value, model_instance)
-
-    def run_validators(self, value: Any) -> None:
-        if isinstance(value, np.ndarray):
-            value = value.tolist()
-        super().run_validators(value)
 
     def formfield(self, form_class: Any = None, choices_form_class: Any = None, **kwargs: Any) -> forms.Field:
         return super().formfield(
@@ -59,18 +51,13 @@ class VectorField(Field):
 
 class VectorWidget(forms.TextInput):
     def format_value(self, value: Any) -> str | None:
-        if isinstance(value, np.ndarray):
-            value = value.tolist()
+        if isinstance(value, Vector):
+            value = value.to_list()
         return super().format_value(value)
 
 
 class VectorFormField(forms.CharField):
     widget = VectorWidget
-
-    def has_changed(self, initial: Any, data: Any) -> bool:
-        if isinstance(initial, np.ndarray):
-            initial = initial.tolist()
-        return super().has_changed(initial, data)
 
     def to_python(self, value: Any) -> Any:
         if isinstance(value, str) and value == '':