+++ tests/test_sqlalchemy.py
@@ -182,10 +182,8 @@ class TestSqlalchemy:
             assert items[0].id % 3 == 1
             assert items[1].id % 3 == 2
             assert items[2].id % 3 == 0
-            assert np.array_equal(items[0].embedding, np.array([1.5, 2, 3]))
-            assert items[0].embedding.dtype == np.float32
-            assert np.array_equal(items[1].embedding, np.array([4, 5, 6]))
-            assert items[1].embedding.dtype == np.float32
+            assert items[0].embedding == Vector([1.5, 2, 3])
+            assert items[1].embedding == Vector([4, 5, 6])
             assert items[2].embedding is None
 
     def test_vector(self, engine):
@@ -193,7 +191,7 @@ class TestSqlalchemy:
             session.add(Item(id=1, embedding=[1, 2, 3]))
             session.commit()
             item = session.get_one(Item, 1)
-            assert np.array_equal(item.embedding, [1, 2, 3])
+            assert item.embedding == Vector([1, 2, 3])
 
     def test_vector_l2_distance(self, engine):
         create_items()
@@ -435,7 +433,7 @@ class TestSqlalchemy:
             session.add(Item(embedding=[1, 2, 3]))
             session.add(Item(embedding=[4, 5, 6]))
             res = session.query(avg(Item.embedding)).one()[0]
-            assert np.array_equal(res, np.array([2.5, 3.5, 4.5]))
+            assert res == Vector([2.5, 3.5, 4.5])
 
     def test_avg_orm(self, engine):
         with Session(engine) as session:
@@ -444,7 +442,7 @@ class TestSqlalchemy:
             session.add(Item(embedding=[1, 2, 3]))
             session.add(Item(embedding=[4, 5, 6]))
             res = session.scalars(select(avg(Item.embedding))).one()
-            assert np.array_equal(res, np.array([2.5, 3.5, 4.5]))
+            assert res == Vector([2.5, 3.5, 4.5])
 
     def test_sum(self, engine):
         with Session(engine) as session:
@@ -453,7 +451,7 @@ class TestSqlalchemy:
             session.add(Item(embedding=[1, 2, 3]))
             session.add(Item(embedding=[4, 5, 6]))
             res = session.query(sum(Item.embedding)).one()[0]
-            assert np.array_equal(res, np.array([5, 7, 9]))
+            assert res == Vector([5, 7, 9])
 
     def test_sum_orm(self, engine):
         with Session(engine) as session:
@@ -462,7 +460,7 @@ class TestSqlalchemy:
             session.add(Item(embedding=[1, 2, 3]))
             session.add(Item(embedding=[4, 5, 6]))
             res = session.scalars(select(sum(Item.embedding))).one()
-            assert np.array_equal(res, np.array([5, 7, 9]))
+            assert res == Vector([5, 7, 9])
 
     def test_bad_dimensions(self, engine):
         item = Item(embedding=[1, 2])
@@ -515,7 +513,7 @@ class TestSqlalchemy:
         with Session(engine) as session:
             session.execute(insert(AutoItem), [{'embedding': np.array([1, 2, 3])}])
             item = session.query(AutoItem).first()
-            assert np.array_equal(item.embedding, [1, 2, 3])
+            assert item.embedding == Vector([1, 2, 3])
 
     def test_half_precision(self, engine):
         create_items()
@@ -563,8 +561,7 @@ class TestSqlalchemyArray:
 
             # this fails if the driver does not cast arrays
             item = session.get_one(Item, 1)
-            assert np.array_equal(item.embeddings[0], [1, 2, 3])
-            assert np.array_equal(item.embeddings[1], [4, 5, 6])
+            assert item.embeddings == [Vector([1, 2, 3]), Vector([4, 5, 6])]
 
     def test_halfvec_array(self, engine):
         with Session(engine) as session:
@@ -587,10 +584,10 @@ class TestSqlalchemyAsync:
 
         async with async_session() as session:
             async with session.begin():
-                embedding = np.array([1, 2, 3])
+                embedding = Vector([1, 2, 3])
                 session.add(Item(id=1, embedding=embedding))
                 item = await session.get_one(Item, 1)
-                assert np.array_equal(item.embedding, embedding)
+                assert item.embedding == embedding
 
         await engine.dispose()
 
@@ -600,10 +597,10 @@ class TestSqlalchemyAsync:
 
         async with async_session() as session:
             async with session.begin():
-                embedding = [1, 2, 3]
+                embedding = HalfVector([1, 2, 3])
                 session.add(Item(id=1, half_embedding=embedding))
                 item = await session.get_one(Item, 1)
-                assert item.half_embedding == HalfVector(embedding)  # ty: ignore[invalid-argument-type]
+                assert item.half_embedding == embedding
 
         await engine.dispose()
 
@@ -631,10 +628,10 @@ class TestSqlalchemyAsync:
 
         async with async_session() as session:
             async with session.begin():
-                embedding = [1, 2, 3]
+                embedding = SparseVector([1, 2, 3])
                 session.add(Item(id=1, sparse_embedding=embedding))
                 item = await session.get_one(Item, 1)
-                assert item.sparse_embedding == SparseVector(embedding)
+                assert item.sparse_embedding == embedding
 
         await engine.dispose()
 
@@ -647,7 +644,7 @@ class TestSqlalchemyAsync:
                 session.add(Item(embedding=[1, 2, 3]))
                 session.add(Item(embedding=[4, 5, 6]))
                 res = await session.scalars(select(avg(Item.embedding)))
-                assert np.array_equal(res.one(), [2.5, 3.5, 4.5])
+                assert res.first() == Vector([2.5, 3.5, 4.5])
 
         await engine.dispose()
 
@@ -665,12 +662,10 @@ class TestSqlalchemyAsyncArray:
             async with session.begin():
                 session.add(Item(id=1, embeddings=[Vector([1, 2, 3]), Vector([4, 5, 6])]))
                 item = await session.get_one(Item, 1)
-                assert np.array_equal(item.embeddings[0], [1, 2, 3])
-                assert np.array_equal(item.embeddings[1], [4, 5, 6])
+                assert item.embeddings == [Vector([1, 2, 3]), Vector([4, 5, 6])]
 
                 session.add(Item(id=2, embeddings=[np.array([1, 2, 3]), np.array([4, 5, 6])]))
                 item = await session.get_one(Item, 2)
-                assert np.array_equal(item.embeddings[0], [1, 2, 3])
-                assert np.array_equal(item.embeddings[1], [4, 5, 6])
+                assert item.embeddings == [Vector([1, 2, 3]), Vector([4, 5, 6])]
 
         await engine.dispose()